NodeGadgetBindings : Export `nodeGadgetMetaclass()`
This is needed by any third-party code using `NodeGadgetClass`. We'd gotten away with it in Gaffer because our only use was in GafferUIModule, where I had mistakenly put the implementation of the function. It is now exported properly from a new GafferUIBindings library.
